backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "autopep8>=2.3.2",
    "boto3>=1.42.89",
    "fastapi>=0.135.3",
    "mangum>=0.21.0",
    "openai>=2.31.0",
    "pypdf>=6.10.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
]
